分裂优惠券表字端优化

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/coupon/CouponListVo.java
@@ -53,15 +53,15 @@ public class CouponListVo {
 		/**
 		 * 分裂优惠卷随机金额最低
 		 */
-		private Integer randomMin;
+		private BigDecimal randomMin;
 		/**
 		 * 分裂优惠卷随机金额最高
 		 */
-		private Integer randomMax;
+		private BigDecimal randomMax;
 		/**
 		 * 分裂优惠券领券人数是否限制 0不限制 1限制
 		 */
-		private Integer receivePerNum;
+		private Byte receivePerNum;
 		/**
 		 * 分裂优惠券可领券人数
 		 */
