+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/coupon/give/CouponGivePopVo.java
@@ -1,9 +1,9 @@
 package com.vpu.mp.service.pojo.shop.coupon.give;
 
+import lombok.Data;
+
 import java.math.BigDecimal;
 import java.sql.Timestamp;
-
-import lombok.Data;
 
 /**
  * 优惠券弹窗
@@ -57,10 +57,10 @@ public class CouponGivePopVo {
 	/**
 	 * 分裂优惠卷随机金额最低
 	 */
-	private Integer randomMin;
+	private BigDecimal randomMin;
 	/**
 	 * 分裂优惠卷随机金额最高
 	 */
-	private Integer randomMax;
+	private BigDecimal randomMax;
 
 }
